final class from AccessControl interface
